JBIDE-13646- Apache Cordova config.xml support
Initial functional version of the config.xml editor support. The big
part of the work is done on the WidgetModel. Now widget model can be
used to sync between a model and the xml document behind. Model objects
implement bean property change listener making it suitable for bindings.
Config.xml editor now has fewer tabs and it is now possible to edit the
model through the GUI.