Support for Android SDK location as a preference
Introduces the Preference support to Hybrid Mobile. There is now a
preference page for Hybrid Mobile. HybridToolsPreferences provides the
access to the individual preferences.

Also introduced a dialog to Android launch shortcuts and launch
configurations to get the SDK location from the users. If a SDK location
is not defined launch does not proceed for Android now.

AndroidSDKManager is modified to use the paths from the preference when
making calls to android command line tools