iOS plugin installation action implementations
Implements the plugin installation actions for iOS. Adds a new library
for making changes on the .pbxproject files. makes some changes to the
general installation actions framework to accomodate values needed on
iOS. Refreshes the plist library to pick up a show stopper bug fix. And
adds a series of tests.