[JBIDE-15779, JBIDE-15770] Run Uninstall action when a featue is removed on the config editor
Introduces the PluginUninstallAction which can be called from context
menu or the config editor. PluginUninstallAction is called when removing
a feature on the config.xml only if a corresponding installed plugin is
found. Otherwise it will do a xml removal only, because features only
carry names this is a possible scenario. 

Also introduces a confirmation dialog for the cordova plugin
uninstallation and changes the progress indicator so that busy indicator
is shown instead of a progress dialog.