Library consistency check
Adds a new function to HybridMobileEngine to check if the underlying
library is compatible with the ProjectGenerator. It goes through every
HybridMobileTemplateResolver supported by the engine and checks if all
the referenced files exists on the library. This is not a full proof
check but it should elliminate most obvious cases of a new engine not
working with the tools.