BUG JBIDE-16213 - Prohibit deployments to AVDs and devices that do not support the minimum SDK version
Enhances the AVD info parsed so that we also have the API level as part
of the AVD information. This introduces the AndroidAVD.java class. Makes
changes to the AndroidLaunchDelegate so that minimum AVD is checked and
a proper exception is thrown. There is also a change on how the default
AVD is selected to account for API levels.

Also modifies the launch configuration UI to check the API level for the
selected AVD