Locator framework
Add a new engine locator framework for searching for engines on the
disk. Changes the implementation of the CordovaEngineProvider so that it
uses the framework to locate the available engines. Introduces
PlatformLibrary to keep track of the locations together with the
HybridMobileEngine instances so that they can be located anywhere.