imported project creation
HybridProjectCreator is modified to be able to create projects without
template files. HybridProjectImportPage now uses the
HybridProjectCreator to create the projects or copy contents to the
created project depending on the options. ConfigJSon can now be created
without engine information. Tests are modified to adjust to API changes
on HybridProjectCreator.