JBIDE-17650 Find better way to enable Tern cordova module for cordova projects
The direct CordovaJS Tern Module configuration for a project is replaced by specifying the Module as Default Module that is to be loaded for
a project with HMT Nature.

This issue depends on: https://github.com/angelozerr/tern.java/pull/87 - Find a way to configure Tern Modules that are default for Nature Adapters defined on a project.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>