JBIDE-17833 Unable to install aerogear cordova push plugin
Many of the cordova plugins do not properly register namespaces on the
plugin.xml which causes failures if parser is namespace aware. This
removes the namespace aware from the plugin.xml parser. As a consequence
we can no longer use XPath to find nodes and this makes a change on how
we retrieve named nodes. 

Also adds several tests for the changes